refactor: 공통되는 스타일 분리
