--- components/src/lib/ConfirmModal/ConfirmModal.styled.ts
@@ -1,9 +1,0 @@
-import styled from "@emotion/styled";
-
-export const ConfirmText = styled.p`
-  color: #0a0d13;
-  font-family: "Noto Sans";
-  font-size: 0.875rem;
-  font-weight: 500;
-  line-height: 0.9375rem;
-`;
