+++ components/src/styles/Typography.styled.ts
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 
-export const AlertText = styled.p`
+export const ModalText = styled.p`
   color: #0a0d13;
   font-family: "Noto Sans";
   font-size: 0.875rem;
